compiler: Use variable pool instead of overwriting a temporary
In certain circumstances we do dumb things like using the first temporary
or overwriting the entire temporary. For code where something like:

    getProperty(s0, $C[1]);

is on the stack, when setting it to a slot, this becomes:

    s0 = obj.types[1];
    obj[obj.slots[1]] = getProperty(s0, $C[1]);

which is obviously clobbered and wrong. So, we need to have some way
to track that temporaries aren't getting clobbered; that is, track which
ones we do and do not use. Thankfully, we already have such a tracker:
the VariablePool. Use this to record temporaries.

Still use the temporary stack when flushing the temporaries out.

This makes crypto-aes.js work when using the compiler.