Relax memoizing getters to enable reinitializing AVM2 on same global
This behavior should probably change in the future. We are not yet borrowing
globals for things as fundamental as Object and Function. So when we apply
instance traits onto Object.prototype, technically those traits should be
non-configurable. This works around how the inspector reinitializes AVM2 upon
every drag and drop, which reloads the builtins and playerGlobal and reapplies
traits. This is not a problem for non-builtins which are not reflected as
native JS builtins, but for things like Object, it tries to reapply a
non-configurable property.