compiler: Support blocks with more than one label
The analyzer was updated to refactor, and added the capacity for a
label case to have more than one label attached to it, but the compiler
was never updated for this case. Refactor the analyzer just a bit more
so that it always creates a list of labels, rather than switching on
a "length" property, update it to handle a case where it forgot that
a label case could have more than one label, and update the compiler
to cope as well.