native: Support istype on constructprop Strings/Numbers
constructprop will always compile into a `new TYPE()`. In any sane type
system, `new String("foo") === "foo"` is an assertion that holds.
Unfortunately, this is JavaScript, so `new String("foo")` is something
that doesn't do what you want out of the box. In order to support this,
we need to make sure we call valueOf() on the types to convert them into
their literal values before checking their intrinsic types.