Compile switches and fix compiling try-catch
Since v8 doesn't support conditional catches, we now only compile one catch
block and do the type tests in an if-else cascade. There's a problem with ASC
generating try blocks with multiple(!) unconditional catch blocks when finally
blocks are used. We ignore all but the first one, because screw ASC.