verifier&compiler: Major verifier refactoring.
The verifier is now using proxy objects for References and Classes instead of the real objects/types retrieved from the domain, however at this point it still retrieves the objects/types from the domain and extracts the traits and name from the real objects to create the proxies. The target is to avoid getting this objects from the domain and to create the proxies from class traits, which are available. The problem is that there is no straightforward way to access base class's traits from a class trait. The only available information about the base class is the super class name (that can be found in classTrait.classInfo.instanceInfo.superName). This class hierarchy can be built statically form the .abc doing a search over all class traits and matching the superName with the class trait name. However this should go through all parent domains and make sure that doesn't break the class shadowing mechanism of AS (a class in a child domain is shadowed by a class with the same name from the the parent domain, and not the other way around).