image: Properly load images with a blank JPEGTables
Some SWFs, like sbemail205.swf (R.I.P.), have a blank JPEGTables.
In this case, we'll accidentally chop off the first two bytes of
the JPEG, which will cause it to fail to load. We shouldn't do that.
This fixes the loading issue with sbemail205.swf, but we don't draw
the bitmaps yet. We'll do that soon.