Added Tabzilla To Top Of The Page.
Steps followed:
1. Added the static tab link (example below) to the top of index.html 
<a href="/en-US/" id="tabzilla">mozilla</a>
2. Included the tabzilla.css CSS file either as a CSS include
<link href="//mozorg.cdn.mozilla.net/media/css/tabzilla-min.css" rel="stylesheet" />
3. Included thetabzilla.js file in template in the footer. 
<script src="//mozorg.cdn.mozilla.net/tabzilla/tabzilla.js"></script>
#Importance
As per says in styleguide:
"Tabzilla is our name for the universal tab that appears on the top of all Mozilla websites (like this one!). It expands to reveal site navigation as well as a promo area (on desktop) and should be a part of any Mozilla-related Web property you’re building. To that end, you’ll find a simple set of HTML/CSS/JS instructions on this page for easy implementation."
link: http://www.mozilla.org/en-US/styleguide/websites/sandstone/tabzilla/