Shape bounding box fixes related to moveTos
A moveTo command mustn't automatically extend bounds, but once a drawing operation starting at its target coordinate happens, the coordinate must be included in the bounds. Additionally, 0,0 isn't automatically part of the bounds, obviously. The Flash player uses 0x8000000 as a sentinel value (which can be seen by getting the bounds of an empty shape in its parent's coordinate space).