Store lineStyle's thickness as twips in coordinates
Turns out the thickness isn't rounded to the nearest pixel after all, but stored - and interpreted - in twips. Because the commands buffer is in bytes, the thickness is written into the coordinates instead. In some ways this is even less dirty than the previous solution. It wastes two bytes, but line styles aren't set that frequently, so this is fine.