Changes to the way constants are embedded in generated code.
I'm trying to hang everything off of the Abc object, which has a reference to the domain it is loaded in. For instance, instead of baking in references to domain objects I now bake in references to MethodInfo objects which in turn have access to the Abc. Abc Info objects have numerical hashes and are baked in as $(112351231), from these hashes I can infer the original Info objects. Even if code was compiled ahead of time.

The remaining fixes here are to deal with bugs there were exposed.