Fixes hitTestPoint on Buttons.
Implements the  `DisplayObject._containsPoint` which works with points in local coordinate spaces.
Keeps a DisplayObject._invertedMatrix in sync with the _matrix as long as no rotation or scaling is applied. This is then used to transform points into the local coordinate space of children when doing recursive containsPoint tests.
Fixes rounding errors in globalToLocal calls.