Only log passing unit tests in verbose mode
... which is introduced as an `--verbose` option to the grunt setup.

Plus, quit with an error in case of failing unit tests.

Without this, failing unit tests are swallowed in a sea of useless spew, making them very easy to miss.

Also restores the ability to specify which unit tests should be run by using the `--tests=path/to/some/js/files` syntax.