Reduces object allocation in the remoting deserializer.
Objects returned by '_readMatrix', '_readRectangle', etc. reuse objects so we must never leak references to them.
GFX frames use getters and setters for the 'matrix' property, setters copy individual members of the matrix so it is safe to assign temporary matrices. Getters return references the internal matrix so the returned value is safe to read but not write.