Optimize some of the cases where BitmapDatas have solid fills.
Flixel uses a BitmapData as a fill texture which is used as a source in copyPixels. The fill texture has a solid color (as a result of fillRect()). If we know that the source texture is a solid color we can usually do better in certain blending / copying operations, so it is worth keeping track of solid fills.