Decode images on the GFX side and defer frame execution until they are loaded.
Add options to the flash package, so that we can trace various API functionality, like Loader for instance which this patch does.
Make GFX and Player side symmetric, in other words they should both be able to send assets to each other. It used to be the case that only the Player could send assets to GFX. All GFX -> Player communication was solely Event messages. This patch generalizes that.