Dispatch loader events after all frame events in the event loop
Because Flash does that, too. To pull this off, we have to also delay committing frames from the parser until the right point of the event loop. For everything but the main SWF's first frame, that is: committing that is what triggers any code execution whatsoever. ... Which we might want to change at some point, too.