Delay all immediately-executed ABC blocks until all frame assets have been resolved
An ABC block can and, in many cases, will rely on other assets being available. It might rely on the fact that an embedded image comes before it in the SWF file, so it can safely just access the image. In delaying the ABC block's execution until all previous assets have become available, we guarantee that that actually works.

(Note: that's not the end of the story, sadly. Code also can rely on just the `bytesLoaded` value matching some expectations, so we have to delay reporting bytes as loaded until their contents are fully available, too. But that's for another bug.)