Bug 1108869 - Part1: delay parsing of loaded data until at least 8kb have been loaded
Well, or the entire file.

This is required to make sure that we've always loaded at least the header of a SWF or image completely, so we don't have to deal with that not being the case in the file type parsers.

More work is needed to properly support SWFs whose first frame is larger than 8kb.