Remove half-baked support for Error#getStackTrace
Because we didn't get it right anyway, and Flash only exposes stack traces in debug builds, so it can't be a compatibility issue. Giving up on this also gives us more breathing room for choosing how we implement builtins, as the details of that can't leak out as easily as with stack traces.