Cache resolved trait function and memoizer in MethodInfo
The same traits are installed in all inheriting classes over and over again. Right now, we also resolve the same functions over and over again, for no good reason whatsoever. With this patch, we instead store them on the trait's `MethodInfo` once and then reuse that cached version, both for the function itself and the memoizer. For lazy compilation-trampolines, each new installation of the trait prototype gets added to the list of getters to patch.