Implement out-of-line exception handling in the interpreter
The interpreter's main loop should be as small as possible, so cluttering it up with type checks and exception handling isn't ideal. OTOH, lots of operations can throw JS exceptions, which we have to turn into AS3 ones before letting content script loose on them. This is achieved by calling a conversion function in the interpreter's `catch` block that takes the current Bytecode and values of various variables and returns a valid exception if it can. Otherwise it at least wraps the exception into a valid one so the internal one doesn't leak.