Round intervals in the shell's task queue to the nearest multiple of 10 instead of fixing them all to MIN_INTERVAL
The fixing to `MIN_INTERVAL` breaks content because it causes events to be delivered in unexpected order. Same for timers: if I set timer `a` to be triggered in 50ms, and timer `b` in 10ms, timer `b` must be guaranteed to fire before `a`, even though it got registered later.