Implement support for arguments.callee
This is quite annoying to support, because it does very different things for free functions compared to traits. Traits are extracted into `MethodClosure`s, including memoization and all that.

Not supported is `arguments.callee` in constructors. It's entirely unclear to me what anybody would use it for: it returns a `MethodClosure` bound to the current instance that can be called and that then executes the constructor's code again on that instance. It can't be `new`'d though. I somewhat doubt that there's code out there that relies on this working, but who knows.