Optimize trait lookups by changing from a 1-dimensional list to a 2-dimensional map setup
Where the names are they keys for the outer dimension and the mangled namespaces those for the inner dimension.

This is made possible by interning namespaces, which means that namespace mangling will happen only once per namespace, so the cost will quickly be amortized.

Cuts trait lookups from ~17% to ~7.5% of total time in Mushroomer.