Add tool for quickly sorting a list of SWF URLs into good/bad lists
Automatically stores state updates in localStorage, so use can be interrupted and continued seamlessly. Also, and importantly, doesn't allow network access or ExternalInterface. We don't support either in the shell, so for the purposes of verifying if something is a good ATS entry, this is useful. And safer than letting these insane SWFs go wild.

`<left arrow>` == good
`<right arrow>` == bad

Alerts when done and logs the list of good files. (Though they can also be found in the textarea with the label "Good").