Perform system.Security.allowDomain for cross-domain SWFs only, instead of image files, too
Currently we delay loading any files with `Loader#load` one event turn and only actually perform the load if the parent SWF has called `system.Security.allowDomain` with a string that allows the loadee's domain. This is to mitigate against a loaded SWF file breaking our sandboxing and reaching into the parent SWF. (Which it is allowed to once `allowDomain` has been called.) The problem is that that prevents loading of image files, too, even though those can't ever perform the attacks we want to prevent.

With this patch, we instead perform the check after the file has started loading, and only if it's a SWF file. In case the check fails, we stop loading and dispatch an asynchronous error event on the `Loader`'s `contentLoaderInfo`.