Improve behavior of PerspectiveProjection to match Flash much more closely
The behavior of `PerspectiveProjection` is interesting in that it can either be completely independent of any `DisplayObject`, or be a live view into one. Creating an instance using `new PerspectiveProjection()` causes the former, getting one via `displayObject.transform.perspectiveProjection` the latter, even though that returns a copy instead of the same instance every time.

This makes the implementation a bit annoying because we need to either store/retrieve the `projectionCenter` and `fieldOfView` values on the instance itself, or on the displayObject it's a view for.