Don't escape AVM1 property names starting with "_"
This was introduced to prevent triggering JS engines' special behavior, such as setting the proto when the property name is `"__proto__"`. That's not required, however: all the special behavior that's there doesn't apply in this case: `__proto__` only has special behavior if used as a literal, not when it's the value of a variable. `__noSuchMethod__`, which only exists in SpiderMonkey, does have special behavior even when contained in a variable. However, that special behavior is only triggered if a missing method is called on the containing object, which we never do on the AVM1 property holders.